Fixed some -Wconversion warnings

+++ test/modules/replacement/lru_collect/lru.cc
@@ -23,7 +23,7 @@ uint32_t CACHE::find_victim(uint32_t triggering_cpu, uint64_t instr_id, uint32_t
   auto end = std::next(begin, NUM_WAY);
 
   // Find the way whose last use cycle is most distant
-  return std::distance(begin, std::min_element(begin, end));
+  return static_cast<uint32_t>(std::distance(begin, std::min_element(begin, end)));
 }
 
 void CACHE::update_replacement_state(uint32_t triggering_cpu, uint32_t set, uint32_t way, uint64_t full_addr, uint64_t ip, uint64_t victim_addr, uint32_t type,
